+++ pybamm/parameters/lithium_ion_parameters.py
@@ -288,10 +288,6 @@ class DomainLithiumIonParameters(BaseParameters):
         # Tortuosity parameters
         self.b_s = self.geo.b_s
 
-        self.C_dl = pybamm.Parameter(
-            f"{Domain} electrode double-layer capacity [F.m-2]"
-        )
-
         # Mechanical parameters
         self.nu = pybamm.Parameter(f"{Domain} electrode Poisson's ratio")
         self.E = pybamm.Parameter(f"{Domain} electrode Young's modulus [Pa]")
@@ -330,6 +326,14 @@ class DomainLithiumIonParameters(BaseParameters):
         )
         self.beta_utilisation = pybamm.Parameter(
             f"{Domain} electrode current-driven interface utilisation factor [m3.mol-1]"
+        )
+
+    def C_dl(self, T):
+        """Dimensional double-layer capacity [F.m-2]"""
+        inputs = {"Temperature [K]": T}
+        Domain = self.domain.capitalize()
+        return pybamm.FunctionParameter(
+            f"{Domain} electrode double-layer capacity [F.m-2]", inputs
         )
 
     def sigma(self, T):